Require DxFeed prop firm selection before credentials
Match the Rithmic Protocol connect flow: pick a prop firm first with no
default selection, then enter credentials. Searchable picker uses a
drawer on mobile and a popover on desktop.
